refactor load8 and load16 declaration

+++ include/pci.hpp
@@ -144,8 +144,8 @@ struct PCI {
   PCIType match(u8*& out_data, u32& offset, u32 addr);
 
   int load32(u32 &val, u32 addr);
-  int load16(u16 &val, u32 addr);
-  int load8(u8 &val, u32 addr);
+  int load16(u32 &val, u32 addr);
+  int load8(u32 &val, u32 addr);
 
   int store32(u32 val, u32 addr);
   int store16(u16 val, u32 addr);
